requirements.txt
streamlit>=1.32.0
openai>=1.14.0
pymupdf>=1.23.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.24
python-dotenv>=1.0.0
